requirements.txt
wandb
numpy
pettingzoo==1.24.2
pytest==7.4.3
pydantic
hydra_core==1.3.2
nbformat>=4.2.0
dash
dash-bootstrap-components
lifelines
statsmodels
dash-mantine-components==0.12.1
python-dotenv
flask_caching
randomname
backoff
openai
pygtrie
regex
setuptools<81
# Dev
black
ipykernel